refactor expiration scheduled tasks
